+++ CHANGES.rst
@@ -24,6 +24,7 @@ v3.0.0 (xxxx-xx-xx)
 - Add ``to_number``.
 - Add ``unquote``.
 - Make ``camel_case`` work with empty string.
+- Make ``capitalize`` only modify the first character of the string. Previously, it would convert rest of string to lower case. (**breaking change**)
 - Make ``splice`` work with strings.